+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version History
 - 2024-??-?? 4.1.21
     - added counter to Analysis/Stats
+    - initial Analysis/Tail draft
     - PixelCalibration example data
 - 2025-01-20 4.1.20
     - Merged LibraryMatching into one graph